third_party/blink/web_tests/fast/events/event-on-xhr-document-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
PASS window.caughtFoo is ['foo@target', 'foo@root']
PASS window.caughtFoo is []
PASS window.caughtFoo is ['foo@target', 'foo@root']
PASS window.caughtFoo is []
PASS successfullyParsed is true

TEST COMPLETE
